3DBlipsSaExclusion "mod"
This is a mod for a mod named "3DBlipsSA". That mod, which is NOT made
by me but by someone named "Xinerki", renders all radar blips on the HUD
at their 3D world positions (together with a distance measurement text).

Someone came to me and asked if I could make it so that the property
icons would be excluded (green house/red house blips), because they can
be omnipresent in some SA:MP servers etc.

This mod "ZZ3DBlipsSaRemoveHouseIcons" does that by hooking into the
loaded mod dll and adds a hook to do an additional check of the blip
type before rendering it.